Accounting clean

+++ Accounting/DTOs/CreateDepreciationDTO.cs
@@ -2,6 +2,6 @@
 {
     public class UpdateDepreciationDTO
     {
-        public double DepreciationAmount { get; set; }
+        public decimal DepreciationAmount { get; set; }
     }
 }
